Modified look of terrain and made a new map
