--- Assets/Gun/Collision.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8e2682f5fac27884692d1c9071a68317
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
